# perf(es/parser): Optimize memory layout