--- a/crates/swc_ecma_parser/src/token.rs
+++ b/crates/swc_ecma_parser/src/token.rs
@@ -126,7 +126,7 @@ pub enum Token {
 
     #[kind(starts_expr)]
     BigInt {
-        value: BigIntValue,
+        value: Box<BigIntValue>,
         raw: Atom,
     },