**Description:**

I boxed `BigIntValue` in `Token` type to make `TokenAndSpan` smaller and made the parser use `Vec` instead of `SmallVec` for token contexts.


## main

```
es/parser/colors        time:   [15.251 us 15.258 us 15.265 us]
es/parser/angular       time:   [7.8481 ms 7.8576 ms 7.8684 ms]
es/parser/backbone      time:   [1.1110 ms 1.1124 ms 1.1143 ms]
es/parser/jquery        time:   [6.1279 ms 6.1357 ms 6.1458 ms]
es/parser/jquery mobile time:   [9.6091 ms 9.6207 ms 9.6346 ms]
es/parser/mootools      time:   [4.6403 ms 4.6483 ms 4.6582 ms]
es/parser/underscore    time:   [949.03 us 949.69 us 950.50 us]
es/parser/three         time:   [27.706 ms 27.747 ms 27.795 ms]
es/parser/yui           time:   [4.6905 ms 4.6970 ms 4.7049 ms]
```


## new

```
es/parser/colors        time:   [14.922 us 14.932 us 14.942 us]
es/parser/angular       time:   [7.7169 ms 7.8253 ms 8.0080 ms]
es/parser/backbone      time:   [1.0840 ms 1.0890 ms 1.0952 ms]
es/parser/jquery        time:   [5.9651 ms 5.9697 ms 5.9754 ms]
es/parser/jquery mobile time:   [9.4014 ms 9.4340 ms 9.4724 ms]
es/parser/mootools      time:   [4.5028 ms 4.5124 ms 4.5255 ms]
es/parser/underscore    time:   [929.05 us 933.09 us 937.61 us]
es/parser/three         time:   [26.885 ms 26.926 ms 26.971 ms]
es/parser/yui           time:   [4.5731 ms 4.5855 ms 4.6022 ms]
```